src/memgraph_import/gemini_client.py
"""
Gemini API client for knowledge graph generation.
"""

import os
from datetime import datetime
from typing import Dict, Any, Optional
from google import genai
from google.genai import types
from .models import NovelChunk, KnowledgeGraph


class GeminiClient:
    """Client for interacting with Google's Gemini API for knowledge graph generation."""

    def __init__(self, api_key: Optional[str] = None):
        """
        Initialize the Gemini client.

        Args:
            api_key: Optional API key. If not provided, will look for GEMINI_API_KEY env var.
        """
        self.api_key = api_key or os.environ.get("GEMINI_API_KEY")
        if not self.api_key:
            raise ValueError(
                "Gemini API key must be provided or set in GEMINI_API_KEY environment variable"
            )

        self.client = genai.Client(api_key=self.api_key)
        self.model = "gemini-2.0-flash-lite"
        self.system_instruction = self._load_system_instruction()

    def _load_system_instruction(self) -> str:
        """Load the system instruction for knowledge graph generation."""
        # This would typically load from a file or constant
        # For now, we'll use a placeholder that should be replaced with the full instruction
        return """
        # Memgraph Literary Knowledge Graph Generator
        
        Generate a single structured JSON file containing both nodes and relationships for Memgraph import.
        This file will contain all entities and their connections in one unified format optimized for literary text analysis.
        
        [Full system instruction would go here - see the notebook for the complete version]
        """

    def generate_knowledge_graph(self, chunk: NovelChunk) -> KnowledgeGraph:
        """
        Generate a knowledge graph from a novel chunk.

        Args:
            chunk: NovelChunk object containing the text to process

        Returns:
            KnowledgeGraph object containing the generated knowledge graph
        """
        contents = [
            types.Content(
                role="user",
                parts=[
                    types.Part.from_text(
                        text=f"""
                        Author: {chunk.author}
                        Book: {chunk.book}
                        Chapter: {chunk.chapter}
                        chunk_order_number: {chunk.chunk_order_number}
                        Chunk: {chunk.chunk}
                        Datetime: {datetime.utcnow().isoformat()}
                        """
                    ),
                ],
            ),
        ]

        generate_content_config = types.GenerateContentConfig(
            response_mime_type="application/json",
            system_instruction=[
                types.Part.from_text(text=self.system_instruction),
            ],
        )

        response = self.client.models.generate_content(
            model=self.model,
            contents=contents,
            config=generate_content_config,
        )

        # Parse the response into a KnowledgeGraph object
        # This is a placeholder - actual implementation would parse the JSON response
        # and convert it into the appropriate model objects
        return self._parse_response(response)

    def _parse_response(self, response: Any) -> KnowledgeGraph:
        """
        Parse the Gemini API response into a KnowledgeGraph object.

        Args:
            response: Raw response from the Gemini API

        Returns:
            KnowledgeGraph object
        """
        # This is a placeholder - actual implementation would parse the JSON response
        # and convert it into the appropriate model objects
        raise NotImplementedError("Response parsing not yet implemented")
