pyproject.toml
[project]
name = "memgraph-import"
version = "0.1.0"
description = "Importing knowledge graphs into Memgraph"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.18.0",
    "ipykernel>=6.29.5",
    "neo4j>=5.28.1",
    "pandas>=2.2.3",
    "pyarrow>=20.0.0",
]
